pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "liftit-cotizador-sicetac"
version = "0.1.0"
description = "MVP FastAPI service that retrieves Sicetac reference tariffs and assists sales executives with trucking quotes."
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
  { name = "Liftit" }
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.25",
  "tenacity>=8.2",
  "python-jose[cryptography]>=3.3",
  "pydantic>=2.5",
  "pydantic-settings>=2.0",
  "defusedxml>=0.7",
  "cachetools>=5.3",
  "sqlalchemy>=2.0",
  "aiofiles>=23.0",
  "gunicorn>=21.2",
  "psycopg2-binary>=2.9",
  "redis>=5.0",
  "sentry-sdk>=1.40"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "httpx-mock>=0.24",
  "ruff>=0.1"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 5050
reload = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "C"]
ignore = ["E501"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
  "tests"
]
asyncio_mode = "auto"
